backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Django Channels (WebSockets)
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0

# Database
mysqlclient>=2.2,<3.0
dj-database-url>=2.1,<3.0

# Utilities
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
gunicorn>=21.2,<23.0
whitenoise>=6.5,<7.0

# Development
django-extensions>=3.2,<4.0
